backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Tutor"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.13.1",
    "anthropic==0.18.1",
    "asyncpg==0.29.0",
    "fastapi==0.109.0",
    "google-generativeai>=0.8.6",
    "langchain>=0.1.0",
    "openai==1.12.0",
    "passlib[bcrypt]==1.7.4",
    "pinecone-client==3.0.3",
    "psycopg2-binary==2.9.9",
    "pydantic==2.6.1",
    "pydantic-settings==2.1.0",
    "email-validator>=2.0.0",
    "pymupdf==1.23.21",
    "python-dotenv==1.0.1",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.2",
    "pandas>=2.2.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "ruff>=0.14.10",
    "rich>=13.0.0",
    "sqlalchemy==2.0.25",
    "tiktoken>=0.12.0",
    "ty>=0.0.10",
    "uvicorn[standard]==0.27.0",
    "google>=3.0.0",
    "pytest>=9.0.2",
    "requests>=2.32.5",
    "google-genai>=1.46.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "pytest-cov>=4.1.0",
    "aiosqlite>=0.19.0",
]
